// tools/early-bound-generator/src/models/interfaces.ts
export interface EbgSettings {
    namespace: string;
    serviceContextName: string;
    outputRelativeDirectory: string;
    builderSettingsJsonRelativePath: string;
    suppressGeneratedCodeAttribute: boolean;
    removeRuntimeVersionComment: boolean;
    suppressAutogeneratedFileHeaderComment: boolean;
    generateTypesAsInternal: boolean;
    includeCommandLine: boolean;
    filePrefixText: string;
    makeReferenceTypesNullable: boolean;
    useDisplayNameForBpfName: boolean;
    addNewFilesToProject: boolean;
    deleteFilesFromOutputFolders: boolean;
    projectNameForEarlyBoundFiles: string;
    tokenCapitalizationOverrides: string[];
    camelCaseClassNames: boolean;
    camelCaseMemberNames: boolean;
    camelCaseCustomWords: string[];
    camelCaseNamesDictionaryRelativePath: string;

    entityTypesFolder: string;
    createOneFilePerEntity: boolean;
    entitiesWhitelist: string[];
    entitiesToSkip: string[];
    entityPrefixesToSkip: string[];
    entityPrefixesWhitelist: string[];
    entityClassNameOverrides: Record<string, string>;
    entityAttributeSpecifiedNames: Record<string, string[]>;
    attributeBlacklist: string[];
    generateAttributeNameConsts: boolean;
    generateAnonymousTypeConstructor: boolean;
    generateConstructorsSansLogicalName: boolean;
    generateEntityRelationships: boolean;
    generateEnumProperties: boolean;
    replaceOptionSetPropertiesWithEnum: boolean;
    generateINotifyPattern: boolean;
    emitVirtualAttributes: boolean;
    emitEntityETC: boolean;
    makeAllFieldsEditable: boolean;
    makeReadonlyFieldsEditable: boolean;
    useLogicalNames: boolean;
    useEnumForStateCodes: boolean;
    useCrmSvcUtilStateEnumNamingConvention: boolean;
    obsoleteDeprecated: boolean;
    obsoleteTokens: string[];
    propertyEnumMappings: string[];
    addDebuggerNonUserCode: boolean;

    optionSetsTypesFolder: string;
    createOneFilePerOptionSet: boolean;
    generateGlobalOptionSets: boolean;
    generateOptionSetMetadataAttribute: boolean;
    addOptionSetMetadataAttribute: boolean;
    generateAllOptionSetLabelMetadata: boolean;
    adjustCasingForEnumOptions: boolean;
    groupLocalOptionSetsByEntity: boolean;
    cleanupCrmSvcUtilLocalOptionSets: boolean;
    invalidCSharpNamePrefix: string;
    localOptionSetFormat: string;
    optionSetLanguageCodeOverride: number | null;
    optionSetNames: Record<string, string>;
    optionNameOverrides: Record<string, string>;
    transliterationRelativePath: string;

    generateMessages: boolean;
    messageTypesFolder: string;
    createOneFilePerMessage: boolean;
    messagesWhitelist: string[];
    messagesToSkip: string[];
    messagePrefixesWhitelist: string[];
    groupMessageRequestWithResponse: boolean;
    generateMessageAttributeNameConsts: boolean;
    makeResponseMessagesEditable: boolean;

    codeCustomizationService: string;
    codeGenerationService: string;
    codeWriterFilterService: string;
    codeWriterMessageFilterService: string;
    metadataProviderService: string;
    metadataQueryProviderService: string;
    namingService: string;

    modelBuilderLogLevel: string;
    updateBuilderSettingsJson: boolean;
    readSerializedMetadata: boolean;
    serializeMetadata: boolean;
    waitForAttachedDebugger: boolean;
}